Implemented most of logging.

+++ sc_server/src/java/no/ebakke/studycaster/backend/Request.java
@@ -6,7 +6,7 @@ public class Request {
   private Long    id;
   private Date    time;
   private String  type;
-  private Integer contentSize;
+  private Long    contentSize;
   private String  remoteAddrHash; // Formerly known as ticketCS
   private String  geoLocation;
   private String  launchTicket; // Formerly known as ticketCC
@@ -16,18 +16,18 @@ public class Request {
   Request() {
   }
 
-  public Request(Date time, String type, Integer contentSize,
+  public Request(Date time, String type, Long contentSize,
       String remoteAddrHash, String geoLocation, String launchTicket,
       String clientCookie, String logEntry)
   {
-    this.time = (Date) time.clone();
-    this.type = type;
-    this.contentSize = contentSize;
+    this.time           = (Date) time.clone();
+    this.type           = type;
+    this.contentSize    = contentSize;
     this.remoteAddrHash = remoteAddrHash;
-    this.geoLocation = geoLocation;
-    this.launchTicket = launchTicket;
-    this.clientCookie = clientCookie;
-    this.logEntry = logEntry;
+    this.geoLocation    = geoLocation;
+    this.launchTicket   = launchTicket;
+    this.clientCookie   = clientCookie;
+    this.logEntry       = logEntry;
   }
 
   public Date getTime() {
@@ -38,7 +38,7 @@ public class Request {
     return type;
   }
 
-  public Integer getContentSize() {
+  public Long getContentSize() {
     return contentSize;
   }
